requirements.txt
python-dotenv==1.0.1
langchain==0.2.0
langchain-groq
streamlit==1.36.0
Pillow==9.5.0
requests==2.31.0
pandas==2.0.3